moved fetchOsmData function to helpers file
